+++ src/components/BuildStatus/styles.js
@@ -1,9 +1,10 @@
 import { StyleSheet } from 'react-native';
 
 const styles = StyleSheet.create({
-  container: {},
-  contentContainer: {},
-  headerBar: { flexDirection: 'row', alignItems: 'center' },
+  container: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
 });
 
 export default styles;