+++ src/config/index.js
@@ -1,7 +1,9 @@
 import googlePlaces from './googlePlaces';
+import version from './version';
 
 const config = {
   googlePlaces,
+  version,
 };
 
 export default config;