+++ src/config/version/index.js
@@ -1,0 +1,9 @@
+const version = {
+  major: 1,
+  minor: 2,
+  patch: 0,
+  build: 9,
+  code: 2,
+};
+
+export default version;