+++ src/scenes/index.js
@@ -1,10 +1,14 @@
 import React from 'react';
 import { Actions, Scene, Tabs, Lightbox } from 'react-native-router-flux';
 
+// Pages
 import Home from './pages/Home';
 import Profile from './pages/Profile';
 import Search from './pages/Search';
 import Place from './pages/Place';
+
+// Modals
+import InfoModal from './modals/InfoModal';
 
 const scenes = Actions.create(
   <Lightbox>
@@ -17,6 +21,8 @@ const scenes = Actions.create(
       <Scene key="search" component={Search} hideNavBar />
       <Scene key="place" component={Place} hideNavBar />
     </Scene>
+
+    <Scene key="infoModal" component={InfoModal} />
   </Lightbox>,
 );